Add whitespace trimming to the #{markdown} tag
This prevents the situation where any indented markdown blocks would
always be rendered as code-blocks due to their initial indentation.

By default all #{markdown} blocks (but not the extensions) will trim
the minimal amount of whitespace necessary to dedent the input. To
disable this feature you can pass the `preserveWhitespace:true` option
to the tag.

For example:

    #{markdown preserveWhitespace:true}
            This whitespace will be preserved.
    #{/markdown}