Add `markdownInline` extension and `#{markdown.inline}` tag
Moved the whitespace processing into the markdown processor itself,
and created an "inline" option for single paragraph output.

    #{markdown.inline}
        A single paragraph
    #{/markdown}

    ${someText.markdownInline()}

Both of those will not produce a paragraph tag; this is useful for if
you're formatting the text which will always be a single paragraph (and
want to use your own tags).